refactor(discord): bootstrap の Minecraft MCP 起動を minecraft-mcp.ts へ分離する
