+++ apps/discord/src/bootstrap.ts
@@ -24,7 +24,6 @@ import { closeDb } from "@vicissitude/store/db";
 import { SqliteMoodStore } from "@vicissitude/store/mood-store";
 import { incrementEmoji } from "@vicissitude/store/queries";
 import { AivisSpeechSynthesizer, createEmotionToTtsStyleMapper } from "@vicissitude/tts";
-import { spawn, type Subprocess } from "bun";
 
 import {
 	createConversationAgentSpecs,
@@ -43,6 +42,7 @@ import {
 } from "./bootstrap/layers.ts";
 import { setupMemoryRecording } from "./bootstrap/memory-recording.ts";
 import { createMetrics } from "./bootstrap/metrics.ts";
+import { startMinecraftMcp } from "./bootstrap/minecraft-mcp.ts";
 import { type AppConfig, loadConfig } from "./config.ts";
 import { DiscordGateway } from "./gateway/discord.ts";
 import {
@@ -54,76 +54,6 @@ import {
 import { MoodNicknameService } from "./mood-nickname.ts";
 import { createPortLayout } from "./port-allocator.ts";
 import { createShutdown } from "./shutdown.ts";
-
-// ─── MCP Process Management ─────────────────────────────────────
-
-async function waitForMcpReady(
-	proc: Subprocess,
-	port: string,
-): Promise<"ready" | "died" | "timeout"> {
-	const processDied = Symbol("died");
-	const exitPromise = proc.exited.then(() => processDied);
-	const maxRetries = 30;
-	/* oxlint-disable no-await-in-loop -- intentional sequential polling */
-	for (let i = 0; i < maxRetries; i++) {
-		const result = await Promise.race([
-			fetch(`http://localhost:${port}/health`)
-				.then((res) => res.status)
-				.catch(() => null),
-			exitPromise,
-		]);
-		if (result === processDied) return "died";
-		if (typeof result === "number" && result >= 200 && result < 300) return "ready";
-		await Bun.sleep(500);
-	}
-	/* oxlint-enable no-await-in-loop */
-	return "timeout";
-}
-
-async function startMinecraftMcp(
-	config: AppConfig,
-	root: string,
-	logger: Logger,
-): Promise<Subprocess | null> {
-	if (!config.minecraft) return null;
-
-	const mcEnv: Record<string, string> = {
-		// 子プロセスに必要な環境変数のみを明示的に渡す
-		PATH: process.env.PATH ?? "",
-		HOME: process.env.HOME ?? "",
-		MC_HOST: config.minecraft.host,
-		MC_PORT: String(config.minecraft.port),
-		MC_USERNAME: config.minecraft.username,
-		MC_AUTH_MODE: config.minecraft.authMode,
-		MC_MCP_PORT: String(config.minecraft.mcpPort),
-		DATA_DIR: resolve(root, "data"),
-	};
-	if (config.minecraft.version) mcEnv.MC_VERSION = config.minecraft.version;
-	if (config.minecraft.profilesFolder) mcEnv.MC_PROFILES_FOLDER = config.minecraft.profilesFolder;
-
-	const mcProcess = spawn({
-		cmd: ["bun", "run", resolve(root, "packages/minecraft/src/server.ts")],
-		env: mcEnv,
-		stdout: "inherit",
-		stderr: "inherit",
-	});
-
-	const port = String(config.minecraft.mcpPort);
-	const status = await waitForMcpReady(mcProcess, port);
-	if (status === "died") {
-		logger.error(`[bootstrap] Minecraft MCP process exited with code ${mcProcess.exitCode}`);
-		return null;
-	}
-	if (status === "timeout") {
-		logger.warn(
-			"[bootstrap] Minecraft MCP server health check timed out, but keeping process alive",
-		);
-		return mcProcess;
-	}
-
-	logger.info("[bootstrap] Minecraft MCP server started");
-	return mcProcess;
-}
 
 // ─── Session Gauge ──────────────────────────────────────────────
